Test showing/hiding a username modal prompt
Currently this is using `ReactTestUtils.findRenderedComponentWithType` and then introspecting the `show` prop on the modal to determine if it's visible. I'd preferably like to use bootstrap classes, but `ReactTestUtils.findRenderedDOMComponentWithClass` doesn't work because the modal is using a [portal](https://github.com/tajo/react-portal) to render the modal outside of the render tree. I'd like to come back to this at some point and see if I can rewrite the tests to avoid introspecting on Modal props.